WIP instagram icon to header and adjusted footer/mobile style w/ @media
