+++ components/Footer.js
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-import Image from 'next/image';
 
 const Container = styled.div`
   display: flex;
@@ -12,9 +11,10 @@ const Container = styled.div`
   bottom: 0;
   width: 100%;
   background-color: #ffffff;
-  /* background: linear-gradient(white 0%, white 15%, #a1bfcd 100%); */
-  /* background-image: linear-gradient(
-135deg,#99b6c6 14%,#bcd8e5 78%); */
+  @media (max-width: 1438px) {
+    position: relative;
+    margin-top: 2rem;
+  }
 `;
 
 const FooterUl = styled.ul`
@@ -37,43 +37,22 @@ const FooterLi = styled.li`
   padding-left: 1rem;
   padding-right: 1rem;
   color: ${(props) => props?.theme?.colors?.oliveDrab72};
-`;
 
-const FooterLink = styled.a`
-  display: flex;
-  align-items: center;
-  padding: 0;
+  @media (max-width: 1438px) {
+    margin: 0.5rem;
+    font-size: 1.2rem;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    text-align: center;
+  }
 `;
-
-const X = 5;
-function InstagramIcon() {
-  return (
-    <Image
-      src="/pics/instagram-brands.svg"
-      alt="instagram icon"
-      width={200 / X}
-      height={200 / X}
-    />
-  );
-}
 
 const Footer = ({ copy }) => {
   const { footer = [] } = copy;
   return (
     <Container>
       <FooterUl>
-        <FooterLi>
-          <FooterLink
-            href="https://www.instagram.com/begoniabaschy/"
-            rel="noreferrer"
-            target="_blank"
-          >
-            {/* credit to https://fontawesome.com/v5.15/icons/instagram?style=brands */}
-            <i className="fab fa-instagram">
-              <InstagramIcon />
-            </i>
-          </FooterLink>
-        </FooterLi>
         {footer.map((item) => {
           return <FooterLi key={item}>{item}</FooterLi>;
         })}
